Adding import code, cleanup, and rest of raw data
This commit checks in all of the state-by-state crime data remaining.
It also includes importation code for all of the states as well as
US totals and gun ownership information. It also cleans up the data
that was imported from the state-by-state .csv files.

Finally, the gun ownership .csv file had some typos that were corrected.
The state names were also changed to be unified with the names used to
import the state-by-state data. This will make it much easier to iterate
over the various states and require us to use only one list of state
names.